+++ SOMcreator/classes.py
@@ -37,17 +37,13 @@ def filterable(func: Callable):
     return inner
 
 
-def get_element_by_uuid(uuid: str, proj: SOMcreator.Project) -> Attribute | PropertySet | Object | Aggregation | None:
-    if uuid is None:
-        return None
-    return proj.get_uuid_dict().get(uuid)
-
 
 class IterRegistry(type):
     _registry = set()
     """ Helper for Iteration"""
 
-    def __iter__(self) -> Iterator[PropertySet | Object | Attribute | Aggregation]:
+    def __iter__(self) -> Iterator[
+        SOMcreator.PropertySet | SOMcreator.Object | SOMcreator.Attribute | SOMcreator.Aggregation]:
         return iter(sorted(list(self._registry), key=lambda x: x.name))
 
     def __len__(self) -> int:
@@ -177,11 +173,12 @@ class Hirarchy(object, metaclass=IterRegistry):
             child.name = value
 
     @property
-    def parent(self) -> PropertySet | Object | Attribute | Aggregation:
+    def parent(self) -> SOMcreator.PropertySet | SOMcreator.Object | SOMcreator.Attribute | SOMcreator.Aggregation:
         return self._parent
 
     @parent.setter
-    def parent(self, parent: PropertySet | Object | Attribute | Aggregation) -> None:
+    def parent(self,
+               parent: SOMcreator.PropertySet | SOMcreator.Object | SOMcreator.Attribute | SOMcreator.Aggregation) -> None:
         if self.parent is not None:
             self.parent._children.remove(self)
         self._parent = parent
@@ -203,14 +200,17 @@ class Hirarchy(object, metaclass=IterRegistry):
             return False
 
     @filterable
-    def get_children(self) -> Iterator[PropertySet | Object | Attribute | Aggregation]:
+    def get_children(self) -> Iterator[
+        SOMcreator.PropertySet | SOMcreator.Object | SOMcreator.Attribute | SOMcreator.Aggregation]:
         return iter(self._children)
 
-    def add_child(self, child: PropertySet | Object | Attribute | Aggregation) -> None:
+    def add_child(self,
+                  child: SOMcreator.PropertySet | SOMcreator.Object | SOMcreator.Attribute | SOMcreator.Aggregation) -> None:
         self._children.add(child)
         child.parent = self
 
-    def remove_child(self, child: PropertySet | Object | Attribute | Aggregation | Hirarchy) -> None:
+    def remove_child(self,
+                     child: SOMcreator.PropertySet | SOMcreator.Object | SOMcreator.Attribute | SOMcreator.Aggregation | Hirarchy) -> None:
         if child in self._children:
             self._children.remove(child)
             child.remove_parent()
@@ -234,204 +234,11 @@ class Hirarchy(object, metaclass=IterRegistry):
         del self
 
 
-class Object(Hirarchy):
-    _registry: set[Object] = set()
-
-    def __init__(self, name: str, ident_attrib: [Attribute, str], uuid: str = None,
-                 ifc_mapping: set[str] | None = None, description: None | str = None,
-                 optional: None | bool = None, abbreviation: None | str = None,
-                 project: None | SOMcreator.Project = None,
-                 filter_matrix: list[list[bool]] = None) -> None:
-        super(Object, self).__init__(name, description, optional, project, filter_matrix)
-        self._registry.add(self)
-        self._property_sets: list[PropertySet] = list()
-        self._ident_attrib = ident_attrib
-        self._aggregations: set[Aggregation] = set()
-        self.custom_attribues = {}
-
-        self._abbreviation = abbreviation
-        if abbreviation is None:
-            self._abbreviation = ""
-
-        self._ifc_mapping = ifc_mapping
-        if ifc_mapping is None:
-            self._ifc_mapping = {"IfcBuildingElementProxy"}
-
-        self.uuid = uuid
-        if uuid is None:
-            self.uuid = str(uuid4())
-
-    def __str__(self):
-        return f"Object {self.name}"
-
-    def __lt__(self, other: Object):
-        return self.ident_value < other.ident_value
-
-    def __copy__(self):
-        new_ident_attribute = None
-        if self.is_concept:
-            ident_pset = None
-            new_ident_attribute = str(self.ident_attrib)
-        else:
-            ident_pset = self.ident_attrib.property_set
-
-        new_property_sets = set()
-        for pset in self.get_property_sets(filter=False):
-            new_pset = cp.copy(pset)
-            new_property_sets.add(new_pset)
-            if pset == ident_pset:
-                new_ident_attribute = new_pset.get_attribute_by_name(self.ident_attrib.name)
-
-        if new_ident_attribute is None:
-            raise ValueError(f"Identifier Attribute could'nt be found")
-
-        new_object = Object(name=self.name, ident_attrib=new_ident_attribute, uuid=str(uuid4()),
-                            ifc_mapping=self.ifc_mapping,
-                            description=self.description, optional=self.is_optional(ignore_hirarchy=True),
-                            abbreviation=self.abbreviation,
-                            project=self.project, filter_matrix=self._filter_matrix)
-
-        for pset in new_property_sets:
-            new_object.add_property_set(pset)
-
-        if self.parent is not None:
-            self.parent.add_child(new_object)
-
-        return new_object
-
-    @property
-    def project(self) -> SOMcreator.Project | None:
-        return self._project
-
-    @property
-    def abbreviation(self) -> str:
-        return self._abbreviation
-
-    @abbreviation.setter
-    def abbreviation(self, value) -> None:
-        self._abbreviation = value
-
-    @property
-    def ifc_mapping(self) -> set[str]:
-        return self._ifc_mapping
-
-    @ifc_mapping.setter
-    def ifc_mapping(self, value: set[str]):
-        value_set = set()
-        for item in value:  # filter empty Inputs
-            if not (item == "" or item is None):
-                value_set.add(item)
-        self._ifc_mapping = value_set
-
-    def add_ifc_map(self, value: str) -> None:
-        self._ifc_mapping.add(value)
-
-    def remove_ifc_map(self, value: str) -> None:
-        self._ifc_mapping.remove(value)
-
-    @property
-    def aggregations(self) -> set[Aggregation]:
-        return self._aggregations
-
-    def add_aggregation(self, node: Aggregation) -> None:
-        self._aggregations.add(node)
-
-    def remove_aggregation(self, node: Aggregation) -> None:
-        self._aggregations.remove(node)
-
-    @property
-    def inherited_property_sets(self) -> dict[Object, list[PropertySet]]:
-        def recursion(recursion_property_sets, recursion_obj: Object):
-            psets = recursion_obj.get_property_sets(filter=False)
-
-            if psets:
-                recursion_property_sets[recursion_obj] = psets
-
-            parent = recursion_obj.parent
-            if parent is not None:
-                recursion_property_sets = recursion(recursion_property_sets, parent)
-            return recursion_property_sets
-
-        property_sets = dict()
-        if self.parent is not None:
-            inherited_property_sets = recursion(property_sets, self.parent)
-        else:
-            inherited_property_sets = dict()
-
-        return inherited_property_sets
-
-    @property
-    def is_concept(self) -> bool:
-        if isinstance(self.ident_attrib, Attribute):
-            return False
-        else:
-            return True
-
-    @property
-    def ident_attrib(self) -> Attribute | str:
-        return self._ident_attrib
-
-    @ident_attrib.setter
-    def ident_attrib(self, value: Attribute) -> None:
-        self._ident_attrib = value
-
-    # override name setter because of intheritance
-    @property
-    def name(self) -> str:
-        return self._name
-
-    @name.setter
-    def name(self, value: str):
-        self._name = value
-
-    def add_property_set(self, property_set: PropertySet) -> None:
-        self._property_sets.append(property_set)
-        property_set.object = self
-
-    def remove_property_set(self, property_set: PropertySet) -> None:
-        if property_set in self._property_sets:
-            self._property_sets.remove(property_set)
-
-    @filterable
-    def get_property_sets(self) -> Iterator[PropertySet]:
-        return iter(self._property_sets)
-
-    @filterable
-    def get_attributes(self, inherit: bool = False) -> Iterator[Attribute]:
-        attributes = list()
-        for property_set in self.get_property_sets(filter=False):
-            attributes += property_set.get_attributes(filter=False)
-        if inherit:
-            attributes += self.parent.get_attributes(inherit=True, filter=False)
-        return iter(attributes)
-
-
-    def delete(self, recursive: bool = False) -> None:
-        super(Object, self).delete(recursive)
-
-        for pset in self.get_property_sets(filter=False):
-            pset.delete(recursive, override_ident_deletion=True)
-
-        for aggregation in self.aggregations.copy():
-            aggregation.delete(recursive)
-
-    def get_property_set_by_name(self, property_set_name: str) -> PropertySet | None:
-        for property_set in self.get_property_sets(filter=False):
-            if property_set.name == property_set_name:
-                return property_set
-        return None
-
-    @property
-    def ident_value(self) -> str:
-        if self.is_concept:
-            return str()
-        return ";".join(str(x) for x in self.ident_attrib.value)
-
 
 class PropertySet(Hirarchy):
     _registry: set[PropertySet] = set()
 
-    def __init__(self, name: str, obj: Object = None, uuid: str = None, description: None | str = None,
+    def __init__(self, name: str, obj: SOMcreator.Object = None, uuid: str = None, description: None | str = None,
                  optional: None | bool = None, project: None | SOMcreator.Project = None,
                  filter_matrix: list[list[bool]] = None) -> None:
         super(PropertySet, self).__init__(name, description, optional, project, filter_matrix)
@@ -516,37 +323,37 @@ class PropertySet(Hirarchy):
             self.object.remove_property_set(self)
 
     @property
-    def object(self) -> Object:
+    def object(self) -> SOMcreator.Object:
         return self._object
 
     @object.setter
-    def object(self, value: Object):
+    def object(self, value: SOMcreator.Object):
         self._object = value
 
     @filterable
-    def get_attributes(self) -> Iterator[Attribute]:
+    def get_attributes(self) -> Iterator[SOMcreator.Attribute]:
         """returns all Attributes even if they don't fit the current Project Phase"""
         return iter(self._attributes)
 
-    def add_attribute(self, value: Attribute) -> None:
+    def add_attribute(self, value: SOMcreator.Attribute) -> None:
         if value.property_set is not None and value.property_set != self:
             value.property_set.remove_attribute(value)
         self._attributes.add(value)
 
         value.property_set = self
         for child in self.get_children(filter=False):
-            attrib: Attribute = cp.copy(value)
+            attrib: SOMcreator.Attribute = cp.copy(value)
             value.add_child(attrib)
             child.add_attribute(attrib)
 
-    def remove_attribute(self, value: Attribute, recursive=False) -> None:
+    def remove_attribute(self, value: SOMcreator.Attribute, recursive=False) -> None:
         if value in self.get_attributes(filter=False):
             self._attributes.remove(value)
             if recursive:
                 for child in value.get_children(filter=False):
                     child.property_set.remove_attribute(child)
         else:
-            logging.warning(f"{self.name} -> {value} not in Attributes")
+            logging.warning(f"{self.name} -> {value} not in SOMcreator.Attributes")
 
     def get_attribute_by_name(self, name: str):
         for attribute in self.get_attributes(filter=False):
@@ -567,7 +374,7 @@ class PropertySet(Hirarchy):
 class Attribute(Hirarchy):
     _registry: set[Attribute] = set()
 
-    def __init__(self, property_set: PropertySet | None, name: str, value: list, value_type: str,
+    def __init__(self, property_set: SOMcreator.PropertySet | None, name: str, value: list, value_type: str,
                  data_type: str = value_constants.LABEL,
                  child_inherits_values: bool = False, uuid: str = None, description: None | str = None,
                  optional: None | bool = None, revit_mapping: None | str = None,
@@ -713,11 +520,11 @@ class Attribute(Hirarchy):
                 child._data_type = value
 
     @property
-    def property_set(self) -> PropertySet:
+    def property_set(self) -> SOMcreator.PropertySet:
         return self._property_set
 
     @property_set.setter
-    def property_set(self, value: PropertySet) -> None:
+    def property_set(self, value: SOMcreator.PropertySet) -> None:
         self._property_set = value
 
     def is_equal(self, attribute: Attribute) -> bool:
@@ -753,7 +560,7 @@ class Aggregation(Hirarchy):
     def __str__(self):
         return self.name
 
-    def __init__(self, obj: Object, parent_connection=value_constants.AGGREGATION, uuid: str | None = None,
+    def __init__(self, obj: SOMcreator.Object, parent_connection=value_constants.AGGREGATION, uuid: str | None = None,
                  description: None | str = None,
                  optional: None | bool = None, filter_matrix: list[list[bool]] = None):
