+++ SOMcreator/datastructure/__init__.py
@@ -1,1 +1,2 @@
 from .project import Project
+from .object_ import Object
