+++ SOMcreator/filehandling/property_set.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 import SOMcreator
-from SOMcreator import classes
 from SOMcreator.filehandling import core
 from SOMcreator.filehandling.constants import PREDEFINED_PSETS, ATTRIBUTES
 from SOMcreator.filehandling import attribute
@@ -11,9 +10,9 @@ if TYPE_CHECKING:
     from SOMcreator import Project
 
 
-def load(proj: Project, pset_dict: PropertySetDict, identifier: str, obj: classes.Object | None) -> None:
+def load(proj: Project, pset_dict: PropertySetDict, identifier: str, obj: SOMcreator.Object | None) -> None:
     name, description, optional, parent, filter_matrix = core.get_basics(proj, pset_dict, identifier)
-    pset = classes.PropertySet(name=name, obj=obj, uuid=identifier, description=description, optional=optional,
+    pset = SOMcreator.PropertySet(name=name, obj=obj, uuid=identifier, description=description, optional=optional,
                                project=proj, filter_matrix=filter_matrix)
     attributes_dict = pset_dict[ATTRIBUTES]
     for ident, attribute_dict in attributes_dict.items():
@@ -23,7 +22,7 @@ def load(proj: Project, pset_dict: PropertySetDict, identifier: str, obj: classe
 
 #### Export ####
 
-def write_entry(pset: classes.PropertySet) -> PropertySetDict:
+def write_entry(pset: SOMcreator.PropertySet) -> PropertySetDict:
     pset_dict: PropertySetDict = dict()
     core.write_basics(pset_dict, pset)
     attributes_dict = dict()
